+++ threes_solver/src/algo/impls/monotones.rs
@@ -15,32 +15,32 @@ impl Algos {
         &self,
         game_state: &GameState,
         filter: Option<&dyn AlgoValueFilter>,
-    ) -> u8 {
+    ) -> f64 {
         assert_not_supported(self, filter);
 
-        let mut score: i8 = 0;
+        let mut score: f64 = 0.0;
 
         iterate_with_neighbors(game_state.get_grid(), |_index, card, neighbors| {
             let right = neighbors[Direction::Right as usize];
             if right < Card::MAX {
                 if right > card {
-                    score += 1;
+                    score += 1.0;
                 } else if right < card {
-                    score -= 1;
+                    score -= 1.0;
                 }
             }
 
             let down = neighbors[Direction::Down as usize];
             if down < Card::MAX {
                 if down > card {
-                    score += 1;
+                    score += 1.0;
                 } else if down < card {
-                    score -= 1;
+                    score -= 1.0;
                 }
             }
         });
 
-        score.abs() as u8
+        score.abs()
     }
 }
 
@@ -57,7 +57,10 @@ mod tests {
     fn test_monotones() {
         for value in 0..=3 {
             let game_state = generate_game_state([value; 16]);
-            assert_eq!(0, Monotones.monotones(&game_state, None), "0 when all cells have the same value {value}");
+            assert_eq!(
+                0.0, Monotones.monotones(&game_state, None),
+                "0 when all cells have the same value {value}"
+            );
         }
 
         let game_state = generate_game_state([
@@ -66,7 +69,7 @@ mod tests {
             1, 2, 3, 6,
             0, 0, 0, 0,
         ]);
-        assert_eq!(3, Monotones.monotones(&game_state, None), "a basic monotone row");
+        assert_eq!(3.0, Monotones.monotones(&game_state, None), "a basic monotone row");
 
         let game_state = generate_game_state([
             0, 0, 0, 0,
@@ -74,7 +77,7 @@ mod tests {
             0, 0, 0, 0,
             0, 0, 0, 0,
         ]);
-        assert_eq!(3, Monotones.monotones(&game_state, None), "it works right to left, too");
+        assert_eq!(3.0, Monotones.monotones(&game_state, None), "it works right to left, too");
 
         let game_state = generate_game_state([
             0, 1, 0, 0,
@@ -82,7 +85,7 @@ mod tests {
             0, 3, 0, 0,
             0, 6, 0, 0,
         ]);
-        assert_eq!(3, Monotones.monotones(&game_state, None), "a basic monotone col");
+        assert_eq!(3.0, Monotones.monotones(&game_state, None), "a basic monotone col");
 
         let game_state = generate_game_state([
             0, 0, 6, 0,
@@ -90,7 +93,7 @@ mod tests {
             0, 0, 2, 0,
             0, 0, 1, 0,
         ]);
-        assert_eq!(3, Monotones.monotones(&game_state, None), "it works down to up, too");
+        assert_eq!(3.0, Monotones.monotones(&game_state, None), "it works down to up, too");
 
         let game_state = generate_game_state([
             6, 3, 2, 1, // -3
@@ -99,7 +102,7 @@ mod tests {
             1, 0, 0, 0, // -1
         // -3 -1 -1 -1
         ]);
-        assert_eq!(6 + 6, Monotones.monotones(&game_state, None), "both directions at the same time");
+        assert_eq!(6.0 + 6.0, Monotones.monotones(&game_state, None), "both directions at the same time");
 
         let game_state = generate_game_state([
             0, 0, 0, 1, // 1
@@ -108,7 +111,7 @@ mod tests {
             1, 2, 3, 6, // 3
         //  1  1  1  3
         ]);
-        assert_eq!(12, Monotones.monotones(&game_state, None), "the other both directions at the same time");
+        assert_eq!(12.0, Monotones.monotones(&game_state, None), "the other both directions at the same time");
 
         let game_state = generate_game_state([
             0, 0, 0, 0,
@@ -116,7 +119,7 @@ mod tests {
             0, 0, 0, 0,
             0, 0, 0, 0,
         ]);
-        assert_eq!(1, Monotones.monotones(&game_state, None), "non-monotones don't score as well");
+        assert_eq!(1.0, Monotones.monotones(&game_state, None), "non-monotones don't score as well");
 
         let game_state = generate_game_state([
             0, 0, 0, 0,
@@ -124,7 +127,10 @@ mod tests {
             0, 0, 0, 0,
             0, 0, 0, 0,
         ]);
-        assert_eq!(1, Monotones.monotones(&game_state, None), "non-monotones don't score as well, the other way");
+        assert_eq!(
+            1.0, Monotones.monotones(&game_state, None),
+            "non-monotones don't score as well, the other way"
+        );
 
         let game_state = generate_game_state([
             0, 3, 2, 0, // 1
@@ -133,7 +139,10 @@ mod tests {
             0, 0, 0, 0, // 0
         //  0  0  1  0
         ]);
-        assert_eq!(3, Monotones.monotones(&game_state, None), "weird behavior in the middle of the board, but that's probably OK in practice");
+        assert_eq!(
+            3.0, Monotones.monotones(&game_state, None),
+            "weird behavior in the middle of the board, but that's probably OK in practice"
+        );
 
         let game_state = generate_game_state([
             6, 3, 2, 1, // 3
@@ -142,7 +151,7 @@ mod tests {
             0, 0, 0, 0, // 0
         //  1  1  1  1
         ]);
-        assert_eq!(6 + 4, Monotones.monotones(&game_state, None), "a pretty-good state");
+        assert_eq!(6.0 + 4.0, Monotones.monotones(&game_state, None), "a pretty-good state");
 
         let game_state = generate_game_state([
             6, 3, 2, 1, // 3
@@ -151,7 +160,10 @@ mod tests {
             0, 0, 0, 0, // 0
         //  2  2  0  0
         ]);
-        assert_eq!(0 + 4, Monotones.monotones(&game_state, None), "a lower but still positive score, with rows alternating directions, which seems reasonable");
+        assert_eq!(
+            0.0 + 4.0, Monotones.monotones(&game_state, None),
+            "a lower but still positive score, with rows alternating directions, which seems reasonable"
+        );
 
         let game_state = generate_game_state([
             6, 3, 2, 1, // 3
@@ -160,7 +172,10 @@ mod tests {
             0, 0, 0, 0, // 0
         //  -1 -1 -1 -1
         ]);
-        assert_eq!(0 + 4, Monotones.monotones(&game_state, None), "interesting to see how it works out with gaps between rows");
+        assert_eq!(
+            0.0 + 4.0, Monotones.monotones(&game_state, None),
+            "interesting to see how it works out with gaps between rows"
+        );
 
         let game_state = generate_game_state([
             48, 24, 12, 6, // 3
@@ -169,7 +184,7 @@ mod tests {
              6,  3,  2, 1, // 3
         //   3   3   3  3
         ]);
-        assert_eq!(3 * 8, Monotones.monotones(&game_state, None), "the best possible state for this algo");
+        assert_eq!(3.0 * 8.0, Monotones.monotones(&game_state, None), "the best possible state for this algo");
 
         let game_state = generate_game_state([
             6, 3, 2, 1, // 3
@@ -178,7 +193,10 @@ mod tests {
             1, 2, 3, 6, // -3
         // -1 -1  1  1
         ]);
-        assert_eq!(0 + 0, Monotones.monotones(&game_state, None), "fully back-and-forth screws you, which is maybe reasonable");
+        assert_eq!(
+            0.0 + 0.0, Monotones.monotones(&game_state, None),
+            "fully back-and-forth screws you, which is maybe reasonable"
+        );
 
         let game_state = generate_game_state([
             24, 12,  2, 3, // -1
@@ -187,6 +205,6 @@ mod tests {
              2,  6,  3, 1, // -1
         //  -2  -1   1 -1
         ]);
-        assert_eq!(4 + 3, Monotones.monotones(&game_state, None), "a complex (dead-end) case");
+        assert_eq!(4.0 + 3.0, Monotones.monotones(&game_state, None), "a complex (dead-end) case");
     }
 }
